refactor(sideloader): unify sideload types and parsing in codexTypes
Consolidates the definition of 'core' extensions and their parsing logic into
a shared location.

Previously, the Sideloader and Conductor had duplicated (and in the Conductor's
case, hard-coded) definitions for core extensions. By moving the SideloadEntry
types and the parseSideloadEntries utility to codexTypes.ts, we establish a
single source of truth for configuration and prepare for the Conductor to
dynamically derive its core extension list from product.json.
