+++ src/stable/src/vs/workbench/contrib/codexConductor/browser/codexTypes.ts
@@ -43,3 +43,29 @@ export function parsePinnedExtensions(value: unknown): PinnedExtensions | undefi
 	}
 	return Object.keys(result).length > 0 ? result : undefined;
 }
+
+/** A string means "install from gallery by ID". An object with `vsix` means "install directly from URL". */
+export interface SideloadVsixEntry {
+	id: string;
+	vsix: string;
+	version: string;
+}
+
+export type SideloadEntry = string | SideloadVsixEntry;
+
+export function parseSideloadEntries(raw: unknown[]): SideloadEntry[] {
+	const entries: SideloadEntry[] = [];
+	for (const item of raw) {
+		if (typeof item === 'string') {
+			entries.push(item);
+		} else if (
+			item && typeof item === 'object' &&
+			typeof (item as Record<string, unknown>).id === 'string' &&
+			typeof (item as Record<string, unknown>).vsix === 'string' &&
+			typeof (item as Record<string, unknown>).version === 'string'
+		) {
+			entries.push(item as SideloadVsixEntry);
+		}
+	}
+	return entries;
+}
